Port smbase to GCC 6.4 with target x86_64-w64-mingw32.
One issue is this platform is compiled with GCC but uses the Microsoft C
run-time library, which is a combination I had not targeted before.

Furthermore, it has 32-bit 'long' but 64-bit pointers, necessitating the
use of stdint.h and uintptr_t; I had previously avoided those due to
portability concerns, but the passage of time has resolve that.

+++ svdict.h
@@ -70,7 +70,7 @@ public:     // types
     string &key() const { return current->key; }
     void *&value() const { return current->value; }
 
-    long private_getCurrent() const { return (long)current; }
+    uintptr_t private_getCurrent() const { return (uintptr_t)current; }
   };
   friend class Iter;
 
@@ -185,7 +185,7 @@ public:
   string toString() const;
 
   // debugging...
-  long private_getTopAddr() const { return (long)top; }
+  uintptr_t private_getTopAddr() const { return (uintptr_t)top; }
 };
 
 #endif // __SVDICT_H
